# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto Trading Bot API for Bybit
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List open positions
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [paper, live, all]
          required: false
      responses:
        "200":
          description: Open positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: openPosition
      tags: [positions]
      summary: Open a manual position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Position opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get a position by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    delete:
      operationId: closePosition
      tags: [positions]
      summary: Close a position
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trade history
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          required: false
        - in: query
          name: mode
          schema:
            type: string
            enum: [paper, live, all]
          required: false
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
...
  /settings/api-keys:
    put:
      operationId: saveApiKeys
      tags: [settings]
      summary: Save Bybit API keys to database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeysInput"
      responses:
        "200":
          description: Save result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyStatus"
    delete:
      operationId: deleteApiKeys
      tags: [settings]
      summary: Remove Bybit API keys from database
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyStatus"
...
components:
  schemas:
...
    Position:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        currentPrice:
          type: number
        quantity:
          type: number
        leverage:
          type: integer
        unrealizedPnl:
          type: number
        unrealizedPnlPercent:
          type: number
        strategy:
          type: string
        mode:
          type: string
          enum: [paper, live]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        openedAt:
          type: string
        bybitOrderId:
          type: ["string", "null"]
        presetName:
          type: ["string", "null"]
      required:
        - id
        - symbol
        - side
        - entryPrice
        - currentPrice
        - quantity
        - leverage
        - unrealizedPnl
        - unrealizedPnlPercent
        - strategy
        - mode
        - openedAt
    PositionInput:
      type: object
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        quantity:
          type: number
        leverage:
          type: integer
        mode:
          type: string
          enum: [paper, live]
        stopLoss:
          type: number
        takeProfit:
          type: number
      required:
        - symbol
        - side
        - quantity
        - leverage
        - mode
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        leverage:
          type: integer
        pnl:
          type: number
        pnlPercent:
          type: number
        strategy:
          type: string
        mode:
          type: string
          enum: [paper, live]
        openedAt:
          type: string
        closedAt:
          type: string
        bybitOrderId:
          type: ["string", "null"]
        wasAveraged:
          type: boolean
        averageCount:
          type: integer
        presetName:
          type: ["string", "null"]
      required:
        - id
        - symbol
        - side
        - entryPrice
        - exitPrice
        - quantity
        - leverage
        - pnl
        - pnlPercent
        - strategy
        - mode
        - wasAveraged
        - averageCount
        - openedAt
        - closedAt
...
    ApiKeysInput:
      type: object
      properties:
        apiKey:
          type: string
        apiSecret:
          type: string
      required:
        - apiKey
        - apiSecret
    ApiKeyStatus:
      type: object
      properties:
        configured:
          type: boolean
        source:
          type: string
          enum: [env, database, none]
        message:
          type: string
      required:
        - configured
        - source
...